Fix buttercup-web inclusion in node
